Reformat data page
